do not delete tests/lock directory when another rt.sh is running

+++ tests/rt.sh
@@ -97,11 +97,13 @@ rt_trap() {
 }
 
 cleanup() {
-  rm -rf ${LOCKDIR}
+  [[ ${delete_lock:-FALSE} == TRUE ]] && rm -rf ${LOCKDIR}
   [[ ${ECFLOW:-false} == true ]] && ecflow_stop
   trap 0
   exit
 }
+
+delete_lock=TRUE
 
 trap '{ echo "rt.sh interrupted"; rt_trap ; }' INT
 trap '{ echo "rt.sh quit"; rt_trap ; }' QUIT
@@ -122,6 +124,7 @@ if mkdir "${LOCKDIR}" ; then
   echo $(hostname) $$ > "${LOCKDIR}/PID"
 else
   echo "Only one instance of rt.sh can be running at a time"
+  delete_lock=FALSE
   exit 1
 fi
